requirements.txt
boto3
botocore
opencv-python
python-dotenv
websocket-client
